requirements.txt
asgiref==3.7.2
attrs==23.1.0
Django==4.2.6
djangorestframework==3.14.0
drf-spectacular==0.26.5
drf-spectacular-sidecar==2023.10.1
inflection==0.5.1
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
mysqlclient==2.1.1
packaging==23.1
pyparsing==3.0.9
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
rpds-py==0.10.6
sqlparse==0.4.4
uritemplate==4.1.1
uwsgi==2.0.22
requests==2.31.0
coverage==7.4.3
botocore==1.34.59
boto3==1.34.58
pymongo==4.7.2